requirements.txt
pip-chill
bentoml
torchvision
joblib
tdqm
wincertstore
from-root
-e .